+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 * Full implementation of `global`
 * `createSandbox`: argument `externalGlobal`
 * Mode the d.ts file to the `typing` directory
+* Clear the code of a JS file (strip BOM and "#!...")
 
 ### 0.0.1 (24.04.2015)